backend/requirements.txt
SQLAlchemy==2.0.48
anthropic==0.86.0
chromadb==1.5.5
fastapi==0.135.1
openpyxl==3.1.5
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
python-docx==1.2.0
python-dotenv==1.2.2
uvicorn==0.42.0
voyageai==0.2.4
